+++ src/datatypes/wrapper/mod.rs
@@ -92,6 +92,7 @@ pub mod first;
 pub mod last;
 pub mod max;
 pub mod memoize;
+pub mod memoizer;
 pub mod min;
 pub mod product;
 pub mod sum;